+++ libnopegl/src/node_block.h
@@ -41,7 +41,5 @@ struct block_info {
 };
 
 void ngli_node_block_extend_usage(struct ngl_node *node, uint32_t usage);
-size_t ngli_node_block_get_cpu_size(struct ngl_node *node);
-size_t ngli_node_block_get_gpu_size(struct ngl_node *node);
 
 #endif